more optimization

+++ nth_sha256.h
@@ -1,6 +1,6 @@
 #ifndef NTH_SHA256_H
 #define NTH_SHA256_H
 
-void nth_sha256(uint8_t digest[], const uint8_t text[], uint32_t length, uint64_t n);
+void nth_sha256(uint8_t digest[], const uint8_t text[], uint64_t length, uint64_t n);
 
 #endif // NTH_SHA256_H
